// src/types/status-enums.ts
export enum TaskStatusEnum {
  Done = 'Done',
  To_Do = 'To_Do',
  On_Hold = 'On_Hold',
  Blocked = 'Blocked',
  Reopened = 'Reopened',
  Canceled = 'Canceled',
  In_Progress = 'In_Progress',
}

export enum ProductionStatusEnum {
  Done = 'Done',
  To_Do = 'To_Do',
  Stopped = 'Stopped',
  Canceled = 'Canceled',
  Launching = 'Launching',
  From_Stock = 'From_Stock',
  In_Progress = 'In_Progress',
}

export enum UserStatusEnum {
  Active = 'Active',
  Pending = 'Pending',
  Inactive = 'Inactive',
}